Rearrange serial I/O code and make 16550 driver generic over T: Io
