+++ src/arch/mod.rs
@@ -1,1 +1,5 @@
+#[cfg(target_arch = "x86_64")]
+#[macro_use]
 pub mod x86_64;
+#[cfg(target_arch = "x86_64")]
+pub use self::x86_64::*;
