+++ src/devices/mod.rs
@@ -1,0 +1,1 @@
+pub mod uart_16550;
